requirements.txt
Django==2.1.5
djangorestframework
markdown
django-filter
psycopg2-binary
django-extensions
kiteconnect
django-model-utils
selenium
requests
django_utils
channels
websocket-client
